cu29_runtime::cutask

Type Alias CuMsg

pub type CuMsg<T> = CuStampedData<T, CuMsgMetadata>;

This is the robotics message type for Copper with the correct Metadata type that will be used by the runtime.

Aliased Type

pub struct CuMsg<T> {
    pub tov: Tov,
    pub metadata: CuMsgMetadata,
    /* private fields */
}

Fields
tov: Tov

The time of validity of the message. It can be undefined (None), one measure point or a range of measures (TimeRange).

metadata: CuMsgMetadata

This metadata is the data that is common to all messages.