cu29_runtime::payload

Struct CuArray

pub struct CuArray<T: Clone, const N: usize> { /* private fields */ }

Copper friendly wrapper for a fixed size array. T: Clone is required because this type derives Reflect, and the reflection path requires the reflected value to be cloneable.

Implementations

impl<T: Clone, const N: usize> CuArray<T, N>

pub fn new() -> Self

pub fn fill_from_iter<I>(&mut self, iter: I)
where I: IntoIterator<Item = T>,

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

pub fn as_slice(&self) -> &[T]

pub fn capacity(&self) -> usize