cu29_runtime::pool
Trait DeviceCuPool

pub trait DeviceCuPool<T: ArrayLike> {
    type O: ArrayLike<Element = T::Element>;

    // Required method
    fn copy_into(
        &self,
        from_device_handle: &CuHandle<Self::O>,
        into_cu_host_memory_pool: &mut CuHostMemoryPool<T>,
    ) -> CuHandle<T>;
}

Required Associated Types

type O: ArrayLike<Element = T::Element>

Required Methods

fn copy_into( &self, from_device_handle: &CuHandle<Self::O>, into_cu_host_memory_pool: &mut CuHostMemoryPool<T>, ) -> CuHandle<T>

Takes a handle to a device buffer and copies it into a host buffer pool. It returns a new handle from the host pool with the data from the device handle given.