cu29_runtime::reflect

Enum Access

pub enum Access<'a> {
    Field(Cow<'a, str>),
    FieldIndex(usize),
    TupleIndex(usize),
    ListIndex(usize),
}

A singular element access within a path. Multiple accesses can be combined into a ParsedPath.

Can be applied to a dyn Reflect to get a reference to the targeted element.

Variants

Field(Cow<'a, str>)

A name-based field access on a struct.

FieldIndex(usize)

A index-based field access on a struct.

TupleIndex(usize)

An index-based access on a tuple.

ListIndex(usize)

An index-based access on a list.

Implementations

impl<'a> Access<'a>

pub fn into_owned(self) -> Access<'static>

Converts this into an “owned” value.

If the Access is of variant Field, the field’s Cow<str> will be converted to its owned counterpart, which doesn’t require a reference.

pub fn display_value(&self) -> &dyn Display

Returns a reference to this Access’s inner value as a &dyn Display.