cu29_runtime::reflect

Struct DynamicEnum

pub struct DynamicEnum { /* private fields */ }

A dynamic representation of an enum.

This allows for enums to be configured at runtime.

Example

// The original enum value
let mut value: Option<usize> = Some(123);

// Create a DynamicEnum to represent the new value
let mut dyn_enum = DynamicEnum::new(
  "None",
  DynamicVariant::Unit
);

// Apply the DynamicEnum as a patch to the original value
value.apply(dyn_enum.as_partial_reflect());

// Tada!
assert_eq!(None, value);

Implementations

impl DynamicEnum

pub fn new<I, V>(variant_name: I, variant: V) -> DynamicEnum
where I: Into<String>, V: Into<DynamicVariant>,

Create a new DynamicEnum to represent an enum at runtime.

Arguments

variant_name: The name of the variant to set

variant: The variant data

pub fn new_with_index<I, V>( variant_index: usize, variant_name: I, variant: V, ) -> DynamicEnum
where I: Into<String>, V: Into<DynamicVariant>,

Create a new DynamicEnum with a variant index to represent an enum at runtime.

Arguments

variant_index: The index of the variant to set

variant_name: The name of the variant to set

variant: The variant data

pub fn set_represented_type( &mut self, represented_type: Option<&'static TypeInfo>, )

Sets the type to be represented by this DynamicEnum.

Panics

Panics if the given type is not a TypeInfo::Enum.

pub fn set_variant<I, V>(&mut self, name: I, variant: V)
where I: Into<String>, V: Into<DynamicVariant>,

Set the current enum variant represented by this struct.

pub fn set_variant_with_index<I, V>( &mut self, variant_index: usize, variant_name: I, variant: V, )
where I: Into<String>, V: Into<DynamicVariant>,

Set the current enum variant represented by this struct along with its variant index.

pub fn variant(&self) -> &DynamicVariant

Get a reference to the DynamicVariant contained in self.

pub fn variant_mut(&mut self) -> &mut DynamicVariant

Get a mutable reference to the DynamicVariant contained in self.

Using the mut reference to switch to a different variant will not update the internal tracking of the variant name and index.

If you want to switch variants, prefer one of the setters: DynamicEnum::set_variant or DynamicEnum::set_variant_with_index.

pub fn from<TEnum>(value: TEnum) -> DynamicEnum
where TEnum: Enum,

Create a DynamicEnum from an existing one.

This is functionally the same as DynamicEnum::from_ref except it takes an owned value.

pub fn from_ref<TEnum>(value: &TEnum) -> DynamicEnum
where TEnum: Enum + ?Sized,

Create a DynamicEnum from an existing one.

This is functionally the same as DynamicEnum::from except it takes a reference.