cu29_runtime::reflect

Struct ParsedPath

pub struct ParsedPath(pub Vec<OffsetAccess>);

A pre-parsed path to an element within a type.

This struct can be constructed manually from its Accesses or with the parse method.

This struct may be used like GetPath but removes the cost of parsing the path string at each element access.

It’s recommended to use this in place of GetPath when the path string is unlikely to be changed and will be accessed repeatedly.

Examples

Parsing a &'static str:

let my_static_string: &'static str = "bar#0.1[2].0";
// Breakdown:
//   "bar" - Access struct field named "bar"
//   "#0" - Access struct field at index 0
//   ".1" - Access tuple struct field at index 1
//   "[2]" - Access list element at index 2
//   ".0" - Access tuple variant field at index 0
let my_path = ParsedPath::parse_static(my_static_string);

Parsing a non-static &str:

let my_string = String::from("bar#0.1[2].0");
// Breakdown:
//   "bar" - Access struct field named "bar"
//   "#0" - Access struct field at index 0
//   ".1" - Access tuple struct field at index 1
//   "[2]" - Access list element at index 2
//   ".0" - Access tuple variant field at index 0
let my_path = ParsedPath::parse(&my_string);

Manually constructing a ParsedPath:

let path_elements = [
    Access::Field(Cow::Borrowed("bar")),
    Access::FieldIndex(0),
    Access::TupleIndex(1),
    Access::ListIndex(2),
    Access::TupleIndex(1),
];
let my_path = ParsedPath::from(path_elements);

Tuple Fields
0: Vec<OffsetAccess>

This is a vector of pre-parsed OffsetAccesses.

Implementations

impl ParsedPath

pub fn parse(string: &str) -> Result<ParsedPath, ReflectPathError<'_>>

Parses a ParsedPath from a string.

Returns an error if the string does not represent a valid path to an element.

The exact format for path strings can be found in the documentation for GetPath. In short, though, a path consists of one or more chained accessor strings. These are:

Named field access (.field)

Unnamed field access (.1)

Field index access (#0)

Sequence access ([2])

Example

#[derive(Reflect)]
struct Foo {
  bar: Bar,
}

#[derive(Reflect)]
struct Bar {
  baz: Baz,
}

#[derive(Reflect)]
struct Baz(f32, Vec<Option<u32>>);

let foo = Foo {
  bar: Bar {
    baz: Baz(3.14, vec![None, None, Some(123)])
  },
};

let parsed_path = ParsedPath::parse("bar#0.1[2].0").unwrap();
// Breakdown:
//   "bar" - Access struct field named "bar"
//   "#0" - Access struct field at index 0
//   ".1" - Access tuple struct field at index 1
//   "[2]" - Access list element at index 2
//   ".0" - Access tuple variant field at index 0

assert_eq!(parsed_path.element::<u32>(&foo).unwrap(), &123);

pub fn parse_static( string: &'static str, ) -> Result<ParsedPath, ReflectPathError<'static>>

Similar to Self::parse but only works on &'static str and does not allocate per named field.