cu29_runtime::reflect

Module tuple_struct

Traits and types used to power tuple-struct-like operations via reflection.

Structs

DynamicTupleStruct

A tuple struct which allows fields to be added at runtime.

TupleStructFieldIter

An iterator over the field values of a tuple struct.

TupleStructInfo

A container for compile-time tuple struct info.

Traits

GetTupleStructField

A convenience trait which combines fetching and downcasting of tuple struct fields.

TupleStruct

A trait used to power tuple struct-like operations via reflection.

Functions

tuple_struct_debug

The default debug formatter for TupleStruct types.

tuple_struct_partial_cmp

Lexicographically compares two TupleStruct values and returns their ordering.

tuple_struct_partial_eq

Compares a TupleStruct with a PartialReflect value.