cu29_runtime
Module monitoring

Some basic internal monitoring tooling Copper uses to monitor itself and the tasks it is running.

Structs

CountingAllocator

A simple allocator that counts the number of bytes allocated and deallocated.

CuDurationStatistics

A Specialized statistics object for CuDuration. It will also keep track of the jitter between the values.

LiveStatistics

Accumulative stat object that can give your some real time statistics.

NoMonitor

A do nothing monitor if no monitor is provided. This is basically defining the default behavior of Copper in case of error.

ScopedAllocCounter

A simple struct that counts the number of bytes allocated and deallocated in a scope.

Enums

CuTaskState

The state of a task.

Decision

Monitor decision to be taken when a task errored out.

Statics

GLOBAL

Traits

CuMonitor

Trait to implement a monitoring task.